+++ sora/config/__init__.py
@@ -1,8 +1,9 @@
 # useful functions
 def test_attr(attr, typ, name):
-    """
-    This function tests if the attribute "attr" belongs to the type "typ",
-    if not it gives an error message informing the name of the variable
+    """Tests attribute.
+
+    This function tests if the attribute ``attr`` belongs to the type ``typ``, 
+    if not it gives an error message informing the name of the variable.
     """
     try:
         return typ(attr)