+++ sora/star/core.py
@@ -1,62 +1,69 @@
-from .utils import search_star, van_belle, kervella, spatial_motion, choice_star
-from .meta import MetaStar
-from astropy.coordinates import SkyCoord, SphericalCosLatDifferential
-from astropy.coordinates import get_sun, SphericalRepresentation, SkyOffsetFrame, ICRS
-from astropy.time import Time
-import astropy.units as u
-import astropy.constants as const
 import warnings
-import numpy as np
-from sora.config import test_attr, input_tests
 
+import astropy.units as u
+import numpy as np
+from astropy.coordinates import SkyCoord
+from astropy.time import Time
+
+from sora.config import test_attr, input_tests
+from sora.config.decorators import deprecated_alias
+from .meta import MetaStar
+from .utils import search_star, van_belle, kervella, spatial_motion, choice_star
 
 warnings.simplefilter('always', UserWarning)
-
 
 __all__ = ['Star']
 
 
 class Star(MetaStar):
+    @deprecated_alias(log='verbose')  # remove this line in v1.0
     def __init__(self, catalogue='gaiaedr3', **kwargs):
-        """ Defines a star
+        """Defines a star.
 
         Parameters:
-            catalogue (str): The catalogue to download data. It can be "gaiadr2" or "gaiaedr3"
+            catalogue (str): The catalogue to download data. It can be 'gaiadr2' 
+              or 'gaiaedr3'.
             code (str): Gaia Source code for searching in VizieR.
-            coord (str, SkyCoord): if code is not given, coord nust have the coordinates
-                RA and DEC of the star to search in VizieR: 'hh mm ss.ss +dd mm ss.ss'
+            coord (str, SkyCoord): if code is not given, coord nust have the 
+              coordinates RA and DEC of the star to search in VizieR: 
+              'hh mm ss.ss +dd mm ss.ss'.
             ra (int, float): Right Ascension, in deg.
             dec (int, float): Declination, in deg.
-            parallax (int, float): Parallax, in mas. Default = 0
-            pmra (int, float): Proper Motion in RA*, in mas/year. Default = 0
-            pmdec (int, float): Proper Motion in DEC, in mas/year. Default = 0
-            rad_vel (int, float): Radial Velocity, in km/s. Default = 0 km/s
-            epoch (str, Time): Epoch of the coordinates. Default = 'J2000'
-            nomad (bool): If true, it tries to download the magnitudes from NOMAD
-                catalogue. Default = True
-            bjones (bool): If true, it uses de star distance from Bailer-Jones et al. (2018)
-                Default = True
-            log (bool): If true, it prints the downloaded information. Default = True
+            parallax (int, float): Parallax, in mas (Default: 0).
+            pmra (int, float): Proper Motion in RA*, in mas/year (Default: 0).
+            pmdec (int, float): Proper Motion in DEC, in mas/year (Default: 0).
+            rad_vel (int, float): Radial Velocity, in km/s (Default: 0).
+            epoch (str, Time): Epoch of the coordinates (Default: 'J2000').
+            nomad (bool): If true, it tries to download the magnitudes from 
+              NOMAD catalogue (Default: True).
+            bjones (bool): If true, it uses de star distance from Bailer-Jones 
+              et al. (2018) (Default = True).
+            verbose (bool): If true, it prints the downloaded information 
+              (Default: True).
             local (bool): If true, it uses the given coordinate in 'coord'
-                as final coordinate. Default = False
+              as final coordinate (Default: False).
 
-        - The user can only give ("coord") or ("ra" and "dec"), but not both.
-        - To download the coordinates from Gaia, "local" must be set as False
+        Note:
+            The user can give either 'coord' or 'ra' and 'dec', but not both.
+        
+            To download the coordinates from Gaia, "local" must be set as False
             and the ("code") or ("coord") or ("ra" and "dec") must be given.
+
             All values downloaded from Gaia will replace the ones given by the user.
+
         """
         self._attributes = {}
         self.mag = {}
         self.errors = {'RA': 0*u.mas, 'DEC': 0*u.mas, 'Plx': 0*u.mas, 'pmRA': 0*u.mas/u.year,
                        'pmDE': 0*u.mas/u.year, 'rad_vel': 0*u.km/u.year}
-        allowed_kwargs = ['bjones', 'code', 'coord', 'dec', 'epoch', 'local', 'log', 'nomad', 'parallax', 'pmdec', 'pmra',
+        allowed_kwargs = ['bjones', 'code', 'coord', 'dec', 'epoch', 'local', 'verbose', 'nomad', 'parallax', 'pmdec', 'pmra',
                           'ra', 'rad_vel']
         input_tests.check_kwargs(kwargs, allowed_kwargs=allowed_kwargs)
         allowed_catalogues = ['gaiadr2', 'gaiaedr3']
         if catalogue not in allowed_catalogues:
             raise ValueError('Catalogue {} is not one of the allowed catalogues {}'.format(catalogue, allowed_catalogues))
         self._catalogue = {'gaiadr2': 'Gaia-DR2', 'gaiaedr3': 'Gaia-EDR3'}[catalogue]
-        self._log = kwargs.get('log', True)
+        self._verbose = kwargs.get('verbose', True)
         local = kwargs.get('local', False)
         self.bjones = False
         if 'code' in kwargs:
@@ -90,15 +97,23 @@ class Star(MetaStar):
             pass
 
     def set_magnitude(self, **kwargs):
-        """ Sets the magnitudes of a star.
+        """Sets the magnitudes of a star.
 
         Parameters:
-            (band name)=(float): The star magnitude for given band. The band name can be any string the user wants.
+            (band name)=(float): The star magnitude for given band. The band 
+              name can be any string the user wants.
 
         Examples:
+            To set the stars magnitude in the band G:
             set_magnitude(G=10)
+            
+            To set the star's magnitude in the band K:
             set_magnitude(K=15)
+            
+            To set the star's magnitude in a customized band:
             set_magnitude(newband=6)
+
+
         """
         for key in kwargs:
             mag = test_attr(kwargs[key], float, key)
@@ -108,48 +123,55 @@ class Star(MetaStar):
             self.mag[key] = mag
 
     def set_diameter(self, diameter):
-        """ Sets an user diameter for the star, in mas.
+        """Sets an user diameter for the star, in mas.
 
         Parameters:
-            diameter (int,float): sets the user diameter of the star, in mas
+            diameter (int,float): sets the user diameter of the star, in mas.
+
         """
         self.diameter_user = diameter*u.mas
 
     def van_belle(self):
-        """ Determines the diameter of a star in mas using equations from van Belle (1999)
-            -- Publi. Astron. Soc. Pacific 111, 1515-1523:
+        """Determines the diameter of a star in mas using equations from van Belle (1999).
+
+        See: Publi. Astron. Soc. Pacific 111, 1515-1523:.
+
         """
         return van_belle(self.mag.get('B'), self.mag.get('V'), self.mag.get('K'))
 
     def kervella(self):
-        """ Determines the diameter of a star in mas using equations from Kervella et. al (2004)
-            -- A&A Vol.  426, No.  1:
+        """Determines the diameter of a star in mas using equations from Kervella et. al (2004).
+        
+        See: A&A Vol.  426, No.  1:.
+
         """
         return kervella(self.mag.get('B'), self.mag.get('V'), self.mag.get('K'))
 
-    def apparent_diameter(self, distance, mode='auto', band='V', star_type='sg', log=True):
-        """ Calculates the apparent diameter of the star at a given distance
+    @deprecated_alias(log='verbose')  # remove this line in v1.0
+    def apparent_diameter(self, distance, mode='auto', band='V', star_type='sg', verbose=True):
+        """Calculates the apparent diameter of the star at a given distance.
 
         Parameters:
-            distance (int, float): Object geocentric distance, in AU
-            mode (str): The mode to calculate the apparent diameter
-                'user': calculates using user given diameter
-                'gaia': calculates using diameter obtained from Gaia
-                'kervella': calculates using Kervella equations
-                'van_belle': calculates using van Belle equations
-                'auto' (default): tries all the above methods until it is able to calculate diameter.
-                    The order of try is the same as shown above (user, Gaia, Kervella, Van Belle).
-            'band' (str): The band filter to calculate the diameter.
-                If mode is 'kervella' or 'van_belle', the filter must be given, 'B' or 'V'.
-                If mode 'auto', 'V' is selected.
-            'star_type' (str): type of star to calculate the diameter.
-                If mode is 'van_belle', the star type must be given.
-                If mode is 'auto', type = 'sg'.
-                Types can be:
-                    - 'sg' for 'Super Giant'
-                    - 'ms' for 'Main Sequence'
-                    - 'vs' for 'Variable Star'
-            'log' (bool): If True, it prints the mode used by 'auto'.
+            distance (int, float): Object geocentric distance, in AU.
+            mode (str): The mode to calculate the apparent diameter.
+              'user': calculates using user given diameter.
+              'gaia': calculates using diameter obtained from Gaia.
+              'kervella': calculates using Kervella equations.
+              'van_belle': calculates using van Belle equations.
+              'auto' (default): tries all the above methods until it is able to calculate diameter.
+              The order of try is the same as shown above (user, Gaia, Kervella, Van Belle).
+            band (str): The band filter to calculate the diameter.
+              If mode is 'kervella' or 'van_belle', the filter must be given, 'B' or 'V'.
+              If mode 'auto', 'V' is selected.
+            star_type (str): type of star to calculate the diameter.
+              If mode is 'van_belle', the star type must be given.
+              If mode is 'auto', type = 'sg'.
+              Accepted types:
+                'sg' for 'Super Giant'.
+                'ms' for 'Main Sequence'.
+                'vs' for 'Variable Star'.
+            verbose (bool): If True, it prints the mode used by 'auto'.
+
         """
         try:
             distance = distance.to(u.km)
@@ -159,7 +181,7 @@ class Star(MetaStar):
         if mode in ['user', 'auto']:
             try:
                 diam = distance*np.tan(self.diameter_user)
-                if log:
+                if verbose:
                     print('Calculating apparent diameter from user defined diameter')
                 return diam.to(u.km)
             except:
@@ -171,7 +193,7 @@ class Star(MetaStar):
         if mode in ['gaia', 'auto']:
             try:
                 diam = distance*np.tan(self.diameter_gaia)
-                if log:
+                if verbose:
                     text = ''
                     if self.bjones:
                         text += ' + Bailer-Jones et al. (2018)'
@@ -190,7 +212,7 @@ class Star(MetaStar):
             diam_kerv = self.kervella().get(band)
             if diam_kerv is None:
                 raise ValueError('Diameter could not be calculated for given band')
-            if log:
+            if verbose:
                 print('Apparent diameter using Kervella et al. (2004)')
             diam = distance*np.tan(diam_kerv)
             return diam.to(u.km)
@@ -205,7 +227,7 @@ class Star(MetaStar):
             diam_van = diam_van.get(band)
             if diam_van is None:
                 raise ValueError('Diameter could not be calculated for given band')
-            if log:
+            if verbose:
                 print('Apparent diameter using van Belle (1999)')
             diam = distance*np.tan(diam_van)
             return diam.to(u.km)
@@ -214,23 +236,25 @@ class Star(MetaStar):
                              "Please define star diameter or B,V,K magnitudes.")
 
     def __searchgaia(self, catalog):
-        """ Searches for the star position in the Gaia catalogue and save informations
+        """Searches for the star position in the Gaia catalogue and save informations.
 
         Parameters:
-            catalog (str): The catalogue to download data. It can be "gaiadr2" or "gaiaedr3"
+            catalog (str): The catalogue to download data. It can be 'gaiadr2' 
+              or 'gaiaedr3'.
+
         """
         catalogues = {'gaiadr2': 'I/345/gaia2', 'gaiaedr3': 'I/350/gaiaedr3'}
         cat = catalogues[catalog]
         if hasattr(self, 'code'):
-            catalogue = search_star(code=self.code, columns=['**'], catalog=cat, log=self._log)
+            catalogue = search_star(code=self.code, columns=['**'], catalog=cat, verbose=self._verbose)
         else:
             catalogue = search_star(coord=self.coord, columns=['**'], radius=2*u.arcsec,
-                                    catalog=cat, log=self._log)
+                                    catalog=cat, verbose=self._verbose)
         if len(catalogue) == 0:
             raise ValueError('No star was found. It does not exist or VizieR is out.')
         catalogue = catalogue[0]
         if len(catalogue) > 1:
-            if self._log:
+            if self._verbose:
                 print('{} stars were found within 2 arcsec from given coordinate.'.format(len(catalogue)))
                 print('The list below is sorted by distance. Please select the correct star')
             catalogue = choice_star(catalogue, self.coord, ['RA_ICRS', 'DE_ICRS', 'Gmag'], source='gaia')
@@ -286,7 +310,7 @@ class Star(MetaStar):
         cov[np.where(np.isnan(cov))] = 0.0
         self.cov = cov
 
-        if self._log:
+        if self._verbose:
             print('1 {} star found G={}'.format(self._catalogue, catalogue['Gmag'][0]))
             print('star coordinate at J{}: RA={} +/- {}, DEC={} +/- {}'.format(self.epoch.jyear,
                   self.ra.to_string(u.hourangle, sep='hms', precision=5), self.errors['RA'],
@@ -297,9 +321,9 @@ class Star(MetaStar):
         """
         columns = ['RAJ2000', 'DEJ2000', 'Bmag', 'Vmag', 'Rmag', 'Jmag', 'Hmag', 'Kmag']
         catalogue = search_star(coord=self.coord, columns=columns, radius=2*u.arcsec,
-                                catalog='I/297/out', log=self._log)
+                                catalog='I/297/out', verbose=self._verbose)
         if len(catalogue) == 0:
-            if self._log:
+            if self._verbose:
                 warnings.warn('No star was found on NOMAD that matches the star')
             return
         catalogue = catalogue[0]
@@ -319,15 +343,19 @@ class Star(MetaStar):
                 errors.append(mag)
                 continue
             self.set_magnitude(**{mag: catalogue[name][0]})
-        if len(errors) > 0 and self._log:
+        if len(errors) > 0 and self._verbose:
             print('Magnitudes in {} were not located in NOMAD'.format(errors))
 
     def geocentric(self, time):
-        """ Calculates the position of the star, propagating the position using parallax and proper motion
+        """Calculates the position of the star, propagating the position using parallax and proper motion.
 
         Parameters:
-            time (float, Time): reference time to apply proper motion and calculate paralax.
+            time (float, Time): reference time to apply proper motion and 
+              calculate paralax.
+
         """
+        from astropy.coordinates import get_sun, SphericalRepresentation, SkyOffsetFrame, ICRS
+
         try:
             time = Time(time)
         except:
@@ -350,10 +378,11 @@ class Star(MetaStar):
         return g_coord
 
     def barycentric(self, time):
-        """ Calculates the position of the star using proper motion
+        """Calculates the position of the star using proper motion.
 
         Parameters:
             time (str, Time): reference time to apply proper motion.
+
         """
         try:
             time = Time(time)
@@ -364,13 +393,14 @@ class Star(MetaStar):
         return n_coord
 
     def error_at(self, time):
-        """ Estimates the star position error at a given time
+        """Estimates the star position error at a given time.
 
         Parameters:
             time (str, Time): reference time to project star error.
 
         Returns:
-            errors in RA* and DEC
+            errors (float): in RA* and DEC.
+
         """
         try:
             time = Time(time)
@@ -382,18 +412,22 @@ class Star(MetaStar):
         return errors[0]*u.mas, errors[1]*u.mas
 
     def add_offset(self, da_cosdec, ddec):
-        """ Adds an offset to the star position
+        """Adds an offset to the star position.
 
         Parameters:
-            da_cosdec (int, float): offset in Delta_alpha_cos_delta in mas
-            ddec (int, float): offset in Delta_delta in mas
+            da_cosdec (int, float): offset in Delta_alpha_cos_delta, in mas.
+            ddec (int, float): offset in Delta_delta, in mas.
+
         """
+        from astropy.coordinates import SphericalCosLatDifferential
+
         dadc = test_attr(da_cosdec, float, 'da_cosdec')
         dd = test_attr(ddec, float, 'ddec')
         self.offset = SphericalCosLatDifferential(dadc*u.mas, dd*u.mas, 0.0*u.km)
 
     def __str__(self):
-        """ String representation of the Star class
+        """String representation of the Star class.
+        
         """
         out = ''
         if hasattr(self, 'code'):